Overhaul of the "number" type rules and internals
Previously we used Go's math/big, which works really well for
arithmetic, but it won't allow us to provide more complex functions like
sqrt(), log(), etc.

The number internals have been rewritten to use a different library,
cockroachdb/apd. While apd will give us the extended functions we will
need it also handles precision in a totally different way. The precision
in apd is akin to "scale" or significant figures. It's not practical to
try and maintain the existing precision rules.

After investigating different options it has also opened a broader
understanding that how the precision worked before was not feasible to
begin with. While great in theory (that all numbers carry their own
precision), it will cause nightmares down the line with constantly
"casting up" to a higher precision whenever needing to do calculations.
Especially since the behavior is also implicit.

Now the language defines a default precision (which now represents
significant figures) for all internal calculations. Provided this value
is large enough (currently at 20) it should not have any adverse affects
on calculation accuracy unless they are dealing with extremely small or
large numbers.

This is more a stop gap solution to provide something practical but
accurate right now. However, in the future this precision either needs
to be configurable (through the CLI or the runtime itself) or a more
dynamic solution needs to be found to allow for those extreme cases
where the precision needs to truly arbitrary at any magnitude.