Added strings packages (#45)
The strings package contains common operations with strings. There are
too many to list here, but they are listed in teh generated README.

There were loads of small fixes and features added to accommodate this
new package. The noteworthy things are:

1. Arrays can be appended with the "+" operators. For example:
`[1, 2] + [3]` produces `[1, 2, 3]`. This works with all types provides
both operands are arrays of the same type. This also extends to assign
operators, for example `a += ["foo"]` to append an element to an
existing array.

2. Comparison of arrays, for example `[1, 2] == [3, 4]`. This is
especially helpful in assertions. The `==` and `!=` support all array
types as long as both operands are the same type.

3. Casting some types, for example `string 1.2300` will cast a number to
a well formatted string, so "1.23". This works for other types as well
to string, but is not exhaustive yet. The syntax looks the same as
creating array and map literals.

4. Negative literals used to require extra processing. Now they are
normalized by the parser into a single literal. This is a nice
performance improvement, but was also required for using negative
numbers in assertions. Also fixed a bug with expressions that start with
a "-" may cause the wrong operator precedence.

5. Starting to move language tests from Go to the language itself. You
can find them in the lib/lang package. This package has no code (only
tests) so it carries no weight in the final binary.

6. Iterating strings is now possible with "for _, c in str".

7. Assignment (or any other statement) is allowed in the next clause of
a for loop. This is probably more of a bug fix rather a new feature.