Improved error handling
There is now a built in Error type and I have provided an example of
using this can custom error types in the README.

When errors are handled you can now use a special "err" variable to
inspect them.

Also found an unrelated bug where the strings package was not being
included in the final ok binary.