Added "ok asm"
The `asm` tool is useful for debugging. It will print out human-readable
descriptions of each of the compiled instructions for one or more
functions. Usage:

    ok asm package-name function1 function2 ...

Without providing any arguments it will use the current package and
output the `main` function, this is equivalent to running:

    ok asm . main

Here is the output from the `hello-world` example:

    func main():
        1 Assign                 # $1 = "Hello, World!"
        2 Print                  # print($1)