Refactor how object states work (#54)
Previously we would allocate a map at the start of the function only if
it was a constructor. Now that map is provided by the VM always, rather
than being allocated sometimes. This makes the instructions simpler, but
also will allow us to inject the state when we call it as a method later
on.

Another important note is that all variables that start with an
uppercase letter will be included in the state (even if it's not a
constructor) everything else (including variable) are considered simple
registers.