Only temporary registers are now stored in "registers" (#58)
All named variables are now stored in the state, including private
variables. This is to allow us to pass a single instance for the parent
scope coming later.

Since there is now private variables in the state we have to be careful
not to expose them through printing as well.