Standard library is now committed to git
Previously, the generated vm/lib.go (containing the compiled standard
library) replied on the Makefile generating it. This means that people
that pulled in the ok package would not have any of the standard library
built.

It's unfortunate that the generated code needs to be committed, so I've
made some cosmetic changes to the rendered code to bring the ~4000 lines
to about 1300 lines. So at least it's more git diff friendly.