Completely redesign compiling and VM
Up until now we could get away with merging all files into a single
structure, then compiling it all in one go, then finally loading it into
the VM and run it.

A more recent addition (the "time" library, coming soon) doesn't work
with this approach because it needs to be able to reference function and
types within it's own library. So just adding the package prefix onto
all functions breaks one side or the other.

This patch totally overhauls how compilation and VM execution works.
Now, packages and compiled independently and stored on disk with
predictable destinations. Then the VM will load these compiled
packages like you would a dynamic library.

This isn't complete, but it is a major stepping stone forward. Not just
to add the needed flexibility for the upcoming "time" package but also a
huge improvement in the compilation and VM toolchain.

Other noteworthy and breaking changes are:

1. The virtual "lang" package contained one production entity, the Error
interface. This has now been merged to an "error" package. It's a little
more annoying to type "error.Error" but I think this makes more sense
for the internals.

2. Package discovery has been improved and made more consistent. Rather
than just paths, we now have the idea of a $OKPATH directory, which
defaults to the current directory. This is where the base of packages
can be found. Then arguments are expected to be package names rather
than relative paths.