Refactor for new "types" packages (#78)
Package types replaces strings are types with messy manipulations and
checks. This patch is large and adds a few more hacks, but doesn't
change any existing behavior.

It is designed to get all the messy stuff out of the way so that we can
start to resolve interface types properly through the compiler (coming
next).