Refactor Call expressions (#79)
Call expressions were overly simplistic in terms of their syntax,
expecting an identifier or identifier-dot-identifier only. This works
for the majority of cases but falls apart when expressions get more
complicated. It also gets in the way of the compiler trying to resolve
imported functions.

Call now wraps an expression (which could be anything that returns a
function).