Packages are now compiled as a single function
Previously, packages were a collection of named function. Anonymous
functions could be within those named function as literals. Now all
functions are literals and the package itself is compiled as an
interface. This makes the VM treat the package import as an instance
rather than the hacky collection of functions it was before.

To minimise the diff it still does a bit of both, but the remaining code
that uses these old structures will be removed in the future.