Fix parent scope variables (#94)
Previously, parent scope variables (eg. ^foo) were not checked in any
way and just hardcoded to be of type number. This was a hack to get the
fundamental mechanics working without all the complexity of adding the
proper implementation to the compiler. Not surprisingly I have now hit
this limit so it has been implemented correctly.

On a side note, there were a bunch of tests commented out that have been
enabled again.