Remove gob encoding from okc
Gob was a nice easy way to store and load the compiled packages, but we
need to replace that part of the tool with JSON before we can implement
a correct linker.

It also has the added bonus of being able read. Although, the JSON
generated right now is very verbose.