Huge refactoring of compiler and VM (#121)
This is perhaps the largest change in ok history. The change started off
with something relatively minor (I don't even know what it was now).
Like pulling on a stray thread the entire sweater had to come apart
before it could be rebuilt.

Many of the changes in this diff have been things that I knew I couldn't
put off any longer. They kind of all had to happen at once. Without
going into too much detail the major points are:

1. An almost complete redesign of many of the compiler and type
resolution internals. This allowed me to remove a few cases of hard
coded types that were just plain wrong but mainly now the type system is
properly aware of imported packages (and therefore types) without weird
hacks and work arounds.

2. All packages are now compiled and merged into a single object for the
VM. Rather than previously having the VM load each package into a
variable at runtime. This is analogous to a linker and a huge leap
forward for compilation and the VM in general.

3. No more huge "vm/lib.go" needing to be committed to git. The
`lib-gen` tool now only packages up the original ok source and presents
a virtual file system to the compiler. So we still get all the benefit
of having standard libraries embedded in the ok binary, but a much
smaller and more sane git footprint.