Fixes stack traces and unhandled errors (#125)
An unhandled error previously resulted in a very unhelpful panic outside
of a test and even worse when originating from inside a test. Now it
prints out the the stack traces for unhandled errors in and out of
tests.

The stack trace itself also has been fixed to use the positions of the
Call operations (and not the surrounding functions).